requirements.txt
Django==5.2
djangorestframework==3.16.0
django-filter==25.1
django-cors-headers==4.7.0
Pillow==11.2.1
coreapi==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0 